Move ParameterGroup class to the SAGESharp.Testing namespace
