+++ SAGESharpTests/SLB/EqualityUtilsTests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using SAGESharp.SLB;
+using SAGESharp.Testing;
 using System.Collections.Generic;
 
 namespace SAGESharpTests.SLB
