+++ SAGESharpTests/SLB/FooterEntryTests.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
+using SAGESharp.Testing;
 using SAGESharpTests;
 
 namespace SAGESharp.SLB
