+++ SAGESharpTests/SLB/IdentifierEqualityTests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using SAGESharp.SLB;
+using SAGESharp.Testing;
 
 namespace SAGESharpTests.SLB
 {
