+++ SAGESharpTests/SLB/Level/Conversation/ConversationFooterGeneratorTests.cs
@@ -1,6 +1,6 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
-using SAGESharpTests;
+using SAGESharp.Testing;
 using System;
 using System.Collections.Generic;
