+++ SAGESharpTests/SLB/Level/Conversation/FrameEqualityTests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using SAGESharp.SLB.Level.Conversation;
+using SAGESharp.Testing;
 using System;
 
 using static SAGESharpTests.SLB.Level.Conversation.Defaults;
