+++ SAGESharpTests/SLB/StringBinaryWriterTests.cs
@@ -1,7 +1,7 @@
 ﻿using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
-using SAGESharpTests;
+using SAGESharp.Testing;
 using System;
 using System.IO;
 using System.Linq;
